more refactoring, added new test for cluster of 3 nodes
These are connected in circle, n1 -> n2 -> n3 -> n1, and replicated sometimes
works and fails other times. Part of the issue seems to be in local-cluster and
part of it seems to be that lemur as app starts a baby gen_server, so there's kind
of a master node. Cleaing out the baobab stores at the beginning of each test helps.